feat: refine chat response with increase max token and refactor serivice folder
