+++ services/__init__.py
@@ -1,10 +1,33 @@
 from .clustering.clusterer import GMMRaptorClusterer
 from .clustering.summarizer import GEMINI_FAST, LLMSummarizer, make_llm
+
+# Core services
+from .core import ChatContextService, ChatService
 from .core.build_tree_service import RaptorBuildService
 from .document.chunk_service import ChunkService
 from .document.persist_document import PersistDocumentCmd, PersistDocumentUseCase
+
+# Embedding services
+from .embedding import EmbeddingService as EmbeddingQueryService
 from .embedding.embedding_service import EmbeddingService
+
+# Provider services
+from .providers import (
+    FPTLLMClient,
+    GeminiChatLLM,
+    ModelRegistry,
+    OpenAIClientAsync,
+    VoyageEmbeddingClientAsync,
+)
 from .providers.embedder_adapter import VoyageEmbedderAdapter
+
+# Retrieval services
+from .retrieval import (
+    AnswerService,
+    QueryRewriteService,
+    RetrievalService,
+    RetrieveBody,
+)
 
 __all__ = [
     "RaptorBuildService",
@@ -17,4 +40,20 @@ __all__ = [
     "GMMRaptorClusterer",
     "PersistDocumentCmd",
     "PersistDocumentUseCase",
+    # Core services
+    "ChatService",
+    "ChatContextService",
+    # Retrieval services
+    "QueryRewriteService",
+    "RetrievalService",
+    "RetrieveBody",
+    "AnswerService",
+    # Embedding services
+    "EmbeddingQueryService",
+    # Provider services
+    "FPTLLMClient",
+    "GeminiChatLLM",
+    "OpenAIClientAsync",
+    "VoyageEmbeddingClientAsync",
+    "ModelRegistry",
 ]
