+++ services/core/__init__.py
@@ -1,1 +1,13 @@
 # Core Business Services
+
+from .chat_context_service import ChatContextService
+from .chat_service import ChatService
+from .chat_session_service import ChatSessionService
+from .message_service import MessageService
+
+__all__ = [
+    "ChatService",
+    "ChatContextService",
+    "ChatSessionService",
+    "MessageService",
+]
