+++ services/embedding/__init__.py
@@ -1,1 +1,9 @@
 # Embedding Services
+
+from .embedding_query_service import EmbeddingService
+from .embedding_service import EmbeddingService as EmbeddingServiceFull
+
+__all__ = [
+    "EmbeddingService",
+    "EmbeddingServiceFull",
+]
