+++ services/providers/__init__.py
@@ -1,1 +1,15 @@
 # LLM and Embedding Provider Services
+
+from .fpt_llm.client import FPTLLMClient
+from .gemini_chat.llm import GeminiChatLLM
+from .model_registry import ModelRegistry
+from .openai_chat.openai_client_async import OpenAIClientAsync
+from .voyage.voyage_client import VoyageEmbeddingClientAsync
+
+__all__ = [
+    "FPTLLMClient",
+    "GeminiChatLLM",
+    "OpenAIClientAsync",
+    "VoyageEmbeddingClientAsync",
+    "ModelRegistry",
+]
