+++ services/retrieval/context_service.py
@@ -1,0 +1,25 @@
+import json
+from typing import Any, Dict, List, Optional
+
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from services.retrieval.prompt_service import PromptService
+
+
+class ContextService:
+    def __init__(self, prompt_service: PromptService):
+        self.prompt_service = prompt_service
+
+    async def summarize_conversation_context(
+        self, session: AsyncSession, session_id: str, max_messages: int = 10
+    ) -> Dict[str, Any]:
+        # This would typically retrieve context from a repository
+        # For now, we'll return a placeholder
+        return {
+            "session_id": session_id,
+            "message_count": 0,
+            "recent_message_count": 0,
+            "topics_discussed": [],
+            "model_used": "unknown",
+            "conversation_length": "short",
+        }
